Refactor cloc-diff-rel and cloc-git-diff-rel-between-tag-branch-commit
This commit refactors the cloc-diff-rel and cloc-git-diff-rel-between-tag-branch-commit files. It removes concurrency from critical calls to get diffs and cloc git-diff-rel command to avoid mixing up the output sent to stdout. Additionally, it adds a summary of all changes and a markdown file as output. The pagedCommand is also refactored to handle empty x_next_page. Finally, the readme is updated and a bug in fetching records is fixed.